tuning icons, putting createlabel in top stack where it should be
