+++ src/screens/FavoritesScreen.tsx
@@ -63,7 +63,7 @@ export const FavoritesScreen = () => {
         tagListType={TagListType.Favorites}
       />
       <FABDown
-        icon={fabOpen ? "minus" : "cog"}
+        icon={fabOpen ? "minus" : "cog-outline"}
         open={fabOpen}
         actions={fabActions}
         onStateChange={({open}) => setFabOpen(open)}
