+++ src/screens/HistoryScreen.tsx
@@ -74,7 +74,7 @@ const HistoryScreen = () => {
         tagListType={TagListType.History}
       />
       <FABDown
-        icon={fabOpen ? "minus" : "cog"}
+        icon={fabOpen ? "minus" : "cog-outline"}
         open={fabOpen}
         actions={fabActions}
         onStateChange={({open}) => setFabOpen(open)}
